Create files for each command
