+++ src/commands/index.js
@@ -1,0 +1,2 @@
+export { default as list } from './list.js'
+export { default as convert } from './convert.js'
