+++ src/commands/list.js
@@ -1,0 +1,12 @@
+import { getAvailableThemes, logArray } from '../theme.js'
+
+const list = async () => {
+    const themes = await getAvailableThemes()
+
+    console.log('Available themes:')
+    logArray(themes.map(theme => theme.name))
+
+    console.log('Run `rbxtheme convert <theme>` with one of the listed themes')
+}
+
+export default list
